server/requirements.txt
azure-functions==1.8.0
requests==2.25.1